+++ sdk/tables/Azure.Data.Tables/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added support for Upsert batch operations.
+- Added support for some numeric type coercion for TableEntity properties.
 
 ## 3.0.0-beta.2 (2020-10-06)